+++ sdk/appcenter-distribute/src/test/java/com/microsoft/appcenter/distribute/download/manager/DownloadManagerRequestTaskTest.java
@@ -55,30 +55,39 @@ public class DownloadManagerRequestTaskTest {
         when(mDownloader.getDownloadManager()).thenReturn(mDownloadManager);
 
         /* Create RequestTask. */
-        mRequestTask = spy(new DownloadManagerRequestTask(mDownloader));
+        mRequestTask = spy(new DownloadManagerRequestTask(mDownloader, "title %1$s (%2$d)"));
         when(mRequestTask.createRequest(any(Uri.class))).thenReturn(mDownloadManagerRequest);
     }
 
     @Test
     public void downloadStarted() {
+        when(mReleaseDetails.getVersion()).thenReturn(1);
+        when(mReleaseDetails.getShortVersion()).thenReturn("1");
         when(mReleaseDetails.isMandatoryUpdate()).thenReturn(false);
 
         /* Perform background task. */
         mRequestTask.doInBackground();
 
         /* Verify. */
+        String expectedTitle = "title 1 (1)";
+        verify(mDownloadManagerRequest).setTitle(eq(expectedTitle));
         verifyZeroInteractions(mDownloadManagerRequest);
         verify(mDownloader).onDownloadStarted(eq(DOWNLOAD_ID), anyLong());
     }
 
     @Test
     public void hideNotificationOnMandatoryUpdate() {
+        when(mReleaseDetails.getVersion()).thenReturn(1);
+        when(mReleaseDetails.getShortVersion()).thenReturn("1");
+        when(mReleaseDetails.isMandatoryUpdate()).thenReturn(false);
         when(mReleaseDetails.isMandatoryUpdate()).thenReturn(true);
 
         /* Perform background task. */
         mRequestTask.doInBackground();
 
         /* Verify. */
+        String expectedTitle = "title 1 (1)";
+        verify(mDownloadManagerRequest).setTitle(eq(expectedTitle));
         verify(mDownloadManagerRequest).setNotificationVisibility(DownloadManager.Request.VISIBILITY_HIDDEN);
         verify(mDownloadManagerRequest).setVisibleInDownloadsUi(false);
         verify(mDownloader).onDownloadStarted(eq(DOWNLOAD_ID), anyLong());